requirements.txt
openai>=1.0.0
pyyaml>=6.0
gradio